ArtifactID: 1823508 Comment: Additional updates to the rexx.h header files.
git-svn-id: file:///local/rexx/oorexx/oorexx.svn@1166 0b6cbdbe-3aab-466e-b73a-abd511dda0a2

+++ sandbox/david/platform/unix/rexxplatformapis.h
@@ -59,9 +59,9 @@ APIRET APIENTRY RexxPullQueue (
         const char *,                          /* Name of queue to read from  */
         PRXSTRING,                             /* RXSTRING to receive data    */
         REXXDATETIME *,                        /* Stor for data date/time     */
-        unsigned long );                       /* wait status (WAIT|NOWAIT)   */
+        unsigned int);                         /* wait status (WAIT|NOWAIT)   */
 typedef APIRET (APIENTRY *PFNREXXPULLQUEUE)(const char *, PCONSTRXSTRING, REXXDATETIME *,
-                                           unsigned long);
+                                           unsigned int);
 
 #endif
