+++ sandbox/david/platform/windows/rexxplatformapis.h
@@ -44,9 +44,9 @@ APIRET APIENTRY RexxPullQueue (
         const char *,                          /* Name of queue to read from  */
         PRXSTRING,                             /* RXSTRING to receive data    */
         SYSTEMTIME *,                          /* Stor for data date/time     */
-        unsigned long );                       /* wait status (WAIT|NOWAIT)   */
+        unsigned int);                         /* wait status (WAIT|NOWAIT)   */
 typedef APIRET (APIENTRY *PFNREXXPULLQUEUE)(const char *, PCONSTRXSTRING, SYSTEMTIME *,
-                                           unsigned long);
+                                           unsigned int);
 
 #endif
